src/generation/generation-requirements.txt
ipykernel
pandas
jsonlines
SPARQLWrapper
nltk
py4j
neo4j
python-dotenv
termcolor
rdflib